FEAT: custom termination conditions on play in turns

+++ clash-of-llms/flask_app/class_api/termination.py
@@ -4,5 +4,5 @@ sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
 class Termination:
     def __init__(self, round, alignment=100):
-        self._round = round
-        self._alignment = alignment
+        self._round = int(round)
+        self._alignment = int(alignment)
